requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.1.2
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.1
h11==0.13.0
idna==3.3
pydantic==1.9.0
PyJWT==2.3.0
pymongo==4.1.0
python-decouple==3.6
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6
